refactored repeated code into location.js
