+++ js/home.js
@@ -1,6 +1,7 @@
 import { getNearbyNotes, getNoteContent, deleteNote } from "./modules/notes.js";
 import { goto } from "./modules/ui.js";
 import { isAdmin, getUser } from "./modules/auth.js";
+import { lastKnownLocation, updateCurrentLocation } from "./modules/location.js";
 
 console.log("home.js is running");
 
@@ -31,77 +32,11 @@ window.updateAdminButtonVisibility = function () {
   }
 };
 
-let lastKnownLocation = {
-  lat: null,
-  lon: null,
-  placeId: null,
-};
-
-async function getPlaceIdFromCoordinates(lat, lon) {
-  return new Promise(async (resolve) => {
-    const center = new google.maps.LatLng(lat, lon);
-
-    const request = {
-      fields: ["displayName", "location", "id"],
-      locationRestriction: {
-        center,
-        radius: 100, // Search within 100 meters
-      },
-      includedPrimaryTypes: [
-        "park",
-        "school",
-        "university",
-        "library",
-        "shopping_mall",
-        "restaurant",
-        "cafe",
-        "bar",
-        "gym",
-        "movie_theater",
-        "museum",
-        "church",
-        "mosque",
-        "synagogue",
-        "city_hall",
-        "police",
-        "post_office",
-      ],
-      maxResultCount: 1,
-      rankPreference: google.maps.places.SearchNearbyRankPreference.DISTANCE,
-    };
-
-    try {
-      const { places } = await google.maps.places.Place.searchNearby(request);
-      if (places && places.length > 0) {
-        resolve(places[0].id);
-      } else {
-        console.warn("⚠️ No nearby places found.");
-        resolve("0");
-      }
-    } catch (error) {
-      console.warn("⚠️ Error during Place.searchNearby:", error);
-      resolve("0");
-    }
-  });
-}
-
 async function updateLocationAndRenderHome() {
   console.log("updateLocationAndRenderHome started.");
   try {
-    const position = await new Promise((resolve, reject) => {
-      navigator.geolocation.getCurrentPosition(resolve, reject);
-    });
-    console.log("Geolocation position obtained:", position);
-
-    lastKnownLocation.lat = position.coords.latitude;
-    lastKnownLocation.lon = position.coords.longitude;
-    console.log("Fetching placeId for lat/lon:", lastKnownLocation.lat, lastKnownLocation.lon);
-    lastKnownLocation.placeId = await getPlaceIdFromCoordinates(
-      lastKnownLocation.lat,
-      lastKnownLocation.lon,
-    );
-    localStorage.setItem("lastKnownLocation", JSON.stringify(lastKnownLocation));
-    console.log("lastKnownLocation updated and stored:", lastKnownLocation);
+    await updateCurrentLocation();
+    console.log("Geolocation position obtained and lastKnownLocation updated:", lastKnownLocation);
 
     console.log("Fetching nearby notes...");
     const notes = await getNearbyNotes(
